+++ modules/core/src-2/AbstractNewtype.scala
@@ -16,27 +16,20 @@
 
 package smithy4s
 
-abstract class Newtype[A] extends AbstractNewtype[A] { self =>
-  opaque type T = A
+abstract class AbstractNewtype[A] extends HasId { self =>
+  // This encoding originally comes from this library:
+  // https://github.com/alexknvl/newtypes#what-does-it-do
+  type Base
+  trait __Tag extends Any
+  type Type <: Base with __Tag
 
-  type Type = T
+  @inline final def value(x: Type): A =
+    x.asInstanceOf[A]
 
-  extension (orig: Type) def value: A = orig
+  def schema: Schema[Type]
 
-  def apply(a: A): Newtype.this.Type = a
-
-  def unapply(orig: Type): Some[A] = Some(orig.value)
-
-  implicit val asBijection: Bijection[A, Type] = new Newtype.Make[A, Type] {
-    def to(a: A): Type = self.apply(a)
-    def from(t: Type): A = value(t)
+  implicit val tag: ShapeTag[Type] = new ShapeTag[Type] {
+    def id: ShapeId = self.id
+    def schema: Schema[Type] = self.schema
   }
-
-  object hint {
-    def unapply(h: Hints): Option[Type] = h.get(tag)
-  }
-}
-
-object Newtype {
-  private[smithy4s] trait Make[A, B] extends Bijection[A, B]
 }